+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+Unreleased
+==================
+
+## Changed:
+
+ * Change `Link` and `Guard` to be public
+
 0.1.0 (2020-07-08)
-==========
+==================
 
 ## Added: